+++ demo/helpers/x-page.js
@@ -1,7 +1,5 @@
 import '@neovici/cosmoz-utils/elements/cz-spinner';
 import '@neovici/cosmoz-viewinfo';
-import '@polymer/iron-icons/editor-icons';
-import '@polymer/iron-icons/iron-icons';
 import '@polymer/paper-button/paper-button';
 import '@polymer/paper-dropdown-menu/paper-dropdown-menu';
 import '@polymer/paper-listbox';
@@ -9,6 +7,7 @@ import '@polymer/paper-toggle-button/paper-toggle-button';
 
 import '../../src/cosmoz-omnitable';
 import './cosmoz-translations';
+import '../cosmoz-omnitable-icon';
 
 import { translatable } from '@neovici/cosmoz-i18next';
 import { html } from '@polymer/polymer/lib/utils/html-tag';
@@ -238,7 +237,7 @@ class XPage extends translatable(PolymerElement) {
 					</template>
 
 					<paper-button slot="actions" on-run="removeItems">
-						<iron-icon icon="delete"></iron-icon>
+						<cosmoz-omnitable-icon icon="delete"></cosmoz-omnitable-icon>
 						<span>Remove [[ selectedItems.length ]] items</span>
 					</paper-button>
 					<paper-item slot="actions"