+++ src/cosmoz-omnitable-column-datetime.js
@@ -1,4 +1,3 @@
-import '@polymer/paper-dropdown-menu/paper-dropdown-menu';
 import '@neovici/cosmoz-datetime-input';
 import './ui-helpers/cosmoz-clear-button';