+++ test/grouped-list.test.js
@@ -3,7 +3,7 @@ import { spy as sinonSpy } from 'sinon';
 import { assert, html, fixture, nextFrame } from '@open-wc/testing';
 import { setupIgnoreWindowResizeObserverLoopErrors } from '@lit-labs/virtualizer/support/resize-observer-errors.js';
 
-import '../src/grouped-list/index.js';
+import '../src/grouped-list/index';
 
 const nextEvent = (element, eventName) =>
 	new Promise((resolve) =>
@@ -28,6 +28,7 @@ suite('empty', () => {
 	setupIgnoreWindowResizeObserverLoopErrors(setup, teardown);
 
 	let element;
+
 	setup(async () => {
 		element = await fixture(basicHtmlFixture);
 		element.data = [];
@@ -198,6 +199,7 @@ suite('empty-groups', () => {
 	setupIgnoreWindowResizeObserverLoopErrors(setup, teardown);
 
 	let element;
+
 	setup(async () => {
 		element = await fixture(basicHtmlFixture);
 		element.data = [